First try in Huffman
The Huffman tree is implemented as an array of the Node class, which has the fields of children, parent, value (symbol), probability of occurrence in the text, and code. After processing and building the tree, the code fields are filled in. It is possible that implementing through an array is not the best idea, and you also need to clear memory... I think that Huffman encoding process is ready.